refactor(web): move chat domain types out of store
